+++ app/src/main/java/com/moe/LiveVisualizer/draw/circle/CircleRadialDraw.java
@@ -41,7 +41,7 @@ public class CircleRadialDraw extends RingDraw
 	}
 
 	@Override
-	public void drawGraph(byte[] buffer, Canvas canvas, final int color_mode,boolean useMode)
+	public void drawGraph(double[] buffer, Canvas canvas, final int color_mode,boolean useMode)
 	{
 		PointF point=getPointF();
 		float radius=getRadius();
