+++ app/src/main/java/com/moe/LiveVisualizer/draw/circle/CircleTriangleDraw.java
@@ -59,7 +59,7 @@ public class CircleTriangleDraw extends RingDraw
 	}
 
 	@Override
-	public void drawGraph(byte[] buffer, Canvas canvas,  final int color_mode,boolean useMode)
+	public void drawGraph(double[] buffer, Canvas canvas,  final int color_mode,boolean useMode)
 	{
 		PointF point=getPointF();
 		float radius=getRadius();
