+++ app/src/main/java/com/moe/LiveVisualizer/draw/circle/RippleDraw.java
@@ -113,7 +113,7 @@ public class RippleDraw extends RingDraw
 	}
 
 	@Override
-	public void drawGraph(byte[] buffer, Canvas canvas, int color_mode, boolean useMode)
+	public void drawGraph(double[] buffer, Canvas canvas, int color_mode, boolean useMode)
 	{
 		if(points==null||points.length!=size())
 			points=new float[size()];
