+++ app/src/main/java/com/moe/LiveVisualizer/draw/line/PopCircleDraw.java
@@ -61,7 +61,7 @@ public class PopCircleDraw extends LineDraw
 	}
 	*/
 	@Override
-	public void drawGraph(byte[] buffer, Canvas canvas, int color_mode, boolean useMode)
+	public void drawGraph(double[] buffer, Canvas canvas, int color_mode, boolean useMode)
 	{
 		Paint paint=getPaint();
 		paint.setStrokeWidth(2);
