+++ app/src/main/java/com/moe/LiveVisualizer/draw/line/SquareDraw.java
@@ -58,7 +58,7 @@ public class SquareDraw extends LineDraw
 		return (int)(getBorderHeight()/getBorderWidth());
 	}
 	@Override
-	public void drawGraph(byte[] buffer, Canvas canvas, int color_mode, boolean useMode)
+	public void drawGraph(double[] buffer, Canvas canvas, int color_mode, boolean useMode)
 	{
 		Paint paint=getPaint();
 		if ( points == null || points.length != size() )
