+++ app/src/main/java/com/moe/LiveVisualizer/draw/line/YamaLineDraw.java
@@ -56,7 +56,7 @@ public class YamaLineDraw extends LineDraw
 
 	}*/
 	@Override
-	public void drawGraph(byte[] buffer, Canvas canvas, int color_mode, boolean useMode)
+	public void drawGraph(double[] buffer, Canvas canvas, int color_mode, boolean useMode)
 	{
 		Paint paint=getPaint();
 		if ( points == null || points.length != size() )
